move compaction config into each harness
Review: compaction is a per-loop policy, so each in-house harness
owns its CompactionConfig and threshold discovery instead of sharing
a top-level config class and a clients module. The duplicated
resolver caches per (base_url, model) and queries the upstream
/models card host-side, since the interception server serves no
/models route for agent-side discovery.
