+++ verifiers/v1/harnesses/bash/harness.py
@@ -1,10 +1,16 @@
 import json
 import os
+import random
 from pathlib import Path
+from typing import Any, cast
+
+from openai import APIError
+from pydantic import PositiveInt, model_validator
+from pydantic_config import BaseConfig
 
 from verifiers.v1.clients import ModelContext
-from verifiers.v1.clients.context import resolve_compaction_threshold
-from verifiers.v1.configs.harness import CompactionConfig, HarnessConfig
+from verifiers.v1.clients.base import build_async_openai
+from verifiers.v1.configs.harness import HarnessConfig
 from verifiers.v1.dialects.chat import message_to_wire
 from verifiers.v1.harness import Harness
 from verifiers.v1.runtimes import ProgramResult, Runtime
@@ -26,6 +32,64 @@ SEARCH_PROMPT = (
     "You also have a search tool that returns Google results (title, URL, snippet) for a query; "
     "use it to research, and use bash (e.g. curl) to read result pages in full when needed."
 )
+
+
+CONTEXT_WINDOW_FIELDS = (
+    "max_model_len",
+    "context_length",
+    "context_window",
+    "max_context_length",
+)
+_context_window_cache: dict[tuple[str, str], int | None] = {}
+
+
+class CompactionConfig(BaseConfig):
+    """Context compaction policy for the bash agent loop."""
+
+    summarize_at_tokens: PositiveInt | tuple[PositiveInt, PositiveInt] | None = None
+    """Compact at this token count. A pair draws a task-seeded threshold. When unset, use
+    90% of the model context window when the provider advertises it."""
+
+    @model_validator(mode="after")
+    def validate_range(self) -> "CompactionConfig":
+        value = self.summarize_at_tokens
+        if isinstance(value, tuple) and value[0] > value[1]:
+            raise ValueError(
+                "`summarize_at_tokens` range must be (lo, hi) with lo <= hi."
+            )
+        return self
+
+    def summarize_threshold(self, task_idx: int | None) -> int | None:
+        value = self.summarize_at_tokens
+        if isinstance(value, tuple):
+            lo, hi = value
+            return random.Random(task_idx or 0).randint(lo, hi)
+        return value
+
+
+async def resolve_compaction_threshold(ctx: ModelContext) -> int | None:
+    """90% of the model context window, when the provider's model card advertises one."""
+    key = (ctx.client.base_url, ctx.model)
+    if key not in _context_window_cache:
+        window = None
+        try:
+            async with build_async_openai(ctx.client) as client:
+                payload = await client.get("/models", cast_to=cast(Any, dict[str, Any]))
+        except APIError:
+            payload = {}
+        for card in payload.get("data") or []:
+            if not isinstance(card, dict) or card.get("id") != ctx.model:
+                continue
+            for field in CONTEXT_WINDOW_FIELDS:
+                value = card.get(field)
+                if isinstance(value, int) and not isinstance(value, bool) and value > 0:
+                    window = value
+                    break
+            break
+        _context_window_cache[key] = window
+
+    window = _context_window_cache[key]
+    return max(1, window * 9 // 10) if window is not None else None
 
 
 class BashHarnessConfig(HarnessConfig):
